Add solution for JavaCollections, task 3707

+++ 4.JavaCollections/src/com/javarush/task/task37/task3707/AmigoSet.java
@@ -74,7 +74,7 @@ public class AmigoSet<E> extends AbstractSet<E> implements Serializable, Cloneab
         outputStream.defaultWriteObject();
 
         outputStream.writeInt(callHiddenMethod(map, "capacity"));
-        outputStream.writeInt(callHiddenMethod(map, "loadFactor"));
+        outputStream.writeFloat(callHiddenMethod(map, "loadFactor"));
 
         outputStream.writeInt(map.size());
         for (E element : map.keySet()) {
